Code review refactoring.
