+++ package.json
@@ -100,6 +100,7 @@
     "mapbox-gl": "^1.13.0",
     "marked": "^2.0.1",
     "mini-css-extract-plugin": "^1.3.9",
+    "node-fetch": "^2.6.1",
     "regression": "^2.0.1",
     "nouislider": "^14.6.4",
     "prettier": "^2.2.1",
